Fix Rust lambda return types and numerus type mapping
- Add return type annotation support to Rust lambda generator
- Fix numerus type mapping: f64 → i64 (per language spec)
- Update test expectations across 12 files

+++ fons/codegen/rs/generator.ts
@@ -61,7 +61,7 @@ import { genScriptumExpression } from './expressions/scriptum';
  */
 const typeMap: Record<string, string> = {
     textus: 'String',
-    numerus: 'f64',
+    numerus: 'i64',
     fractus: 'f64',
     magnus: 'i128',
     bivalens: 'bool',
